Fixes LiveSync
Fixes LiveSync by restarting session. Each time we got terminate sessions event we wait 10 sec to get new port. If we get new port for debugging we restart the session.

Move cli process to main process because each restart the session is created(but the tns process is not killed) and we don't have reference to the long running tns process. Also conceptual it's better to have the cli output in the extension output but not in the debug console.

Fixes Debug Console messages and show exactly in which file is the console log.