fix: debugging is not working when app path is changed
Debugging is not working when project has `nsconfig.json` and the path to `app` directory is changed. The problem is that on device the directory is always called `app`, but NativeScript projects can rename it locally to `src` for example.
In order to fix the issue, check if there's nsconfig.json file and replace the `app` from the device path with the local name of the app directory.

This fixes debugging for new Angular templates.