Fix debugger hangs if build of app fails
We wait for port to attach from the CLI and if the port is null we stop the attach process. If the CLI exit with non zero code we send empty port so the debugger is stopped.
Now stop debugging each time CLI exits when it's in initial start state.