fix: use strict versions for all dependencies
Currently build of the extension fails as the versions of dependencies and devDependencies are not hardcoded and we receive version of `@types/lodash` that is incompatible with the TypeScript we are using.
For `@types/lodash` we have the following dist-tags:
```
{ latest: '4.14.122',
  'ts2.0': '4.14.50',
  'ts2.1': '4.14.56',
  'ts2.2': '4.14.112',
  'ts2.3': '4.14.112',
  'ts2.4': '4.14.112',
  'ts2.5': '4.14.112',
  'ts2.6': '4.14.121',
  'ts2.7': '4.14.121',
  'ts2.8': '4.14.122',
  'ts2.9': '4.14.122',
  'ts3.0': '4.14.122',
  'ts3.1': '4.14.122',
  'ts3.2': '4.14.122',
  'ts3.3': '4.14.122',
  'ts3.4': '4.14.122' }
```

In our case we are using TypeScript 2.6.2, so set the version of `@types/lodash` to 4.14.121.
Also set strict versions for all other dependencies (based on what has been installed in my local `node_modules` dir).