fix: debug-brk does not work with webpack
In case you are using `--bundle`/`--hmr` with `--debug-brk`, application should break in `starter.js`. However, an error is shown. There are several problems - first of app the projectName is calculated incorrectly for iOS, so the path to `platforms/ios/<project name>` is incorrect.
Also, as we are modifying the `relativePath` string multiple times, it contains the dir from `nsconfig.json` instead of `app`, but in `platforms` dir, where `starter.js` is located, there's only `app` dir (i.e. no `src` there). Refactor the code, so we'll check the location of the file correctly. Add unit tests for iOS as well.