fix: handle webpack source maps when library is set
In case the `library` path is set in webpack.config.js file, the URLs in sourceMaps are generated in a way that is not expected for our parser - currently we expect `webpack:///<path>`, while it becomes `webpack://<library>/<path>`.
To handle this, add the correct sourceMapPathOverrides in case the user had not specified it already.

Add unit tests and add as devDependency `@types/sinon`. However, the project cannot be transpiled with the current TypeScript version (2.6) when this package is installed, so update TypeScript to latest version.
Use proxyquire in the tests to mock the require of the webpack.config.js file.